More fixes and cleaning
